app/requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.4
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.24.0
attrs==21.2.0
backcall==0.2.0
backpack==0.1
blinker==1.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
cleo==0.6.8
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
databases==0.5.3
dateparser==1.1.0
decorator==5.1.0
Faker==0.8.18
greenlet==1.1.2
idna==3.3
inflection==0.3.1
ipython==7.28.0
jedi==0.18.0
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
loguru==0.5.3
matplotlib==3.4.3
matplotlib-inline==0.1.3
multidict==5.2.0
numpy==1.21.3
orator==0.9.9
parso==0.8.2
pastel==0.1.1
pendulum==1.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
ptyprocess==0.7.0
pyaml==16.12.2
Pygments==2.10.0
pylev==1.4.0
pyparsing==2.4.7
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
regex==2021.10.8
requests==2.26.0
rich==10.12.0
simplejson==3.17.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.25
text-unidecode==1.2
traitlets==5.1.0
typesystem==0.3.1
typing-extensions==3.10.0.2
tzlocal==1.5.1
ujson==4.2.0
urllib3==1.26.7
wcwidth==0.2.5
websockets==9.1
wrapt==1.13.2
yarl==1.7.0
